script: Move script modules logic into subfolder
This moves the two files related to script modules into
a dedicated subfolder. It also splits up the import map
logic from `script_module`, since that file is doing too
many different things.

Testing: it compiles
