+++ components/script/dom/workers/sharedworkerglobalscope.rs
@@ -57,7 +57,7 @@ use crate::dom::types::DebuggerGlobalScope;
 use crate::dom::webgpu::identityhub::IdentityHub;
 use crate::dom::workerglobalscope::WorkerGlobalScope;
 use crate::messaging::{CommonScriptMsg, ScriptEventLoopReceiver, ScriptEventLoopSender};
-use crate::script_module::fetch_a_module_script_graph;
+use crate::modules::script_module::fetch_a_module_script_graph;
 use crate::script_runtime::ScriptThreadEventCategory::WorkerEvent;
 use crate::script_runtime::{IntroductionType, Runtime};
 use crate::task_queue::{QueuedTask, QueuedTaskConversion, TaskQueue};
