+++ components/script/dom/workers/workerglobalscope.rs
@@ -90,9 +90,9 @@ use crate::dom::workernavigator::WorkerNavigator;
 use crate::fetch::{CspViolationsProcessor, Fetch, RequestWithGlobalScope, load_whole_resource};
 use crate::messaging::{CommonScriptMsg, ScriptEventLoopReceiver, ScriptEventLoopSender};
 use crate::microtask::{MicrotaskQueue, MicrotaskRunnable, UserMicrotask};
+use crate::modules::script_module::ScriptFetchOptions;
 use crate::network_listener::{FetchResponseListener, ResourceTimingListener, submit_timing};
 use crate::realms::enter_auto_realm;
-use crate::script_module::ScriptFetchOptions;
 use crate::script_runtime::{IntroductionType, Runtime, get_reports};
 use crate::task::TaskCanceller;
 use crate::task_manager::TaskManager;
